chore(clippy): auto-fix machine-applicable clippy lints (pass 2, full workspace)
Second `cargo clippy --fix` sweep with deny-by-default lints downgraded to
warnings during the pass so the fixer traverses the entire workspace instead of
aborting at the first hard-error crate. Applies machine-applicable rewrites for
approx_constant (literal -> std::consts), absurd_extreme_comparisons, and the
long tail of mechanical lints across all crates and test targets. Remaining
warnings need manual judgment and are handled in follow-up commits.

+++ code/packages/rust/image-codec-jxl/src/entropy.rs
@@ -285,7 +285,7 @@ mod tests {
 
     #[test]
     fn residuals_mixed() {
-        let r: Vec<i32> = (0..50).map(|i| i as i32 - 25).collect();
+        let r: Vec<i32> = (0..50).map(|i| i - 25).collect();
         rt_residuals(&r);
     }
